requirements.txt
torch>=2.0
openai-whisper>=20230314
sounddevice>=0.4.6
scipy>=1.10